+++ src/test/java/seedu/address/model/remark/RemarkTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.property;
+package seedu.address.model.remark;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
